edgeR Differential Expression Analysis

Performs differential expression analysis on a formatted DESeq2 object.

Usage

edgeR_diffexp(
  dds_object,
  condition_col,
  ref_group_name,
  init_pval_cutoff = 1,
  adjust.method = "bonferroni"
)

Arguments

dds_object

A formatted DESeq2 object where each sample and its counts are aligned with the respective metadata in @colData.

condition_col

The name of the condition column in the metadata table in colData.

ref_group_name

The name of the control group to compare against for each comparison. For example, "healthy" or "control".

init_pval_cutoff

The initial p-value cutoff used for filtering.

adjust.method

choose a method of pval correct Family Wise Error Rate reccomended

Value

A list of data frames containing every possible combination of binary comparisons (control vs. experimental group) within your metadata.

Details

Performs differential expression analysis on a formatted DESeq2 object.

INPUTS:

A DESeq2 object with metadata columns corresponding to the samples supplied.

condition: A string giving the name of the condition column within the DESeq2 object's @colData metadata.

reference_group_name: A string specifying the control group for the experiment.

OUTPUTS:

A list of data frames containing each possible combination of binary comparisons (control vs. experimental group) within the supplied metadata.

The resulting data frames contain the following columns:

log2FC: The log2-scaled fold change of a given gene across the specified samples.

P-value: The probability of obtaining data as extreme as the observed values assuming that the null hypothesis is true.

Q-value (FDR): The proportion of false positives after accounting for multiple testing.